fix tests; fix rational, regenerate goldens (bench and code)
