+++ plutarch-test/goldens/api.ctx.uplc.golden
@@ -2,6 +2,6 @@ term (program 1.0.0 #d8799fd8799f9fd8799fd8799fd8799f41a0ff00ffd8799fd8799fd87a9
 get.txInfo (program 1.0.0 ((\i0 -> force headList (force (force sndPair) (unConstrData i1))) #d8799fd8799f9fd8799fd8799fd8799f41a0ff00ffd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffff80a0a141c0a149736f6d65746f6b656e018080d8799fd8799fd87980d87a80ffd8799fd87b80d87a80ffff9f45ab01fe235c4312301443abcdefff80d8799f41b0ffffd87a9fd8799fd8799f41a0ff00ffffff))
 get.mint (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> i3 ((\i0 -> i3 (i3 (i3 i1))) (i4 i1))) ((\i0 -> i3 (i4 i1)) #d8799fd8799f9fd8799fd8799fd8799f41a0ff00ffd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffff80a0a141c0a149736f6d65746f6b656e018080d8799fd8799fd87980d87a80ffd8799fd87b80d87a80ffff9f45ab01fe235c4312301443abcdefff80d8799f41b0ffffd87a9fd8799fd8799f41a0ff00ffffff)) (force tailList)) (force headList)) (\i0 -> i2 (unConstrData i1))) (force (force sndPair))))
 get.credentials (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> i2 (\i0 -> i2 i2 i1)) (\i0 -> i2 (\i0 -> i2 i2 i1))) (\i0 -> \i0 -> force (force (force chooseList) i1 (delay [  ]) (delay (force mkCons (i3 (i5 i1)) (i2 (i4 i1))))))) (\i0 -> i3 (i5 (unConstrData ((\i0 -> i4 (i5 i1)) ((\i0 -> i4 (i5 i1)) (i3 (i2 (i4 i1)))))))) ((\i0 -> unListData (i3 (i4 i1))) ((\i0 -> i3 (i4 i1)) #d8799fd8799f9fd8799fd8799fd8799f41a0ff00ffd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffff80a0a141c0a149736f6d65746f6b656e018080d8799fd8799fd87980d87a80ffd8799fd87b80d87a80ffff9f45ab01fe235c4312301443abcdefff80d8799f41b0ffffd87a9fd8799fd8799f41a0ff00ffffff))) (force tailList)) (force headList)) (\i0 -> i2 (unConstrData i1))) (force (force sndPair))))
-get.sym (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> unBData ((\i0 -> i7 (i10 i1)) ((\i0 -> (\i0 -> i3 (\i0 -> i2 (unMapData i1) (\i0 -> \i0 -> \i0 -> i2 (mapData i3)) (delay (\i0 -> \i0 -> force i1)))) (\i0 -> (\i0 -> force (i10 (force nullList i1) (delay (\i0 -> \i0 -> force i1)) (delay (\i0 -> \i0 -> i2 i3)))) (i3 (\i0 -> force (i10 (equalsData i1 (iData 0)) (delay (\i0 -> \i0 -> force i1)) (delay (\i0 -> \i0 -> i2 i3)))) i1)) i1) (unMapData ((\i0 -> i10 ((\i0 -> i10 (i10 (i10 i1))) (i11 i1))) ((\i0 -> i10 (i11 i1)) #d8799fd8799f9fd8799fd8799fd8799f41a0ff00ffd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffff80a0a141c0a149736f6d65746f6b656e018080d8799fd8799fd87980d87a80ffd8799fd87b80d87a80ffff9f45ab01fe235c4312301443abcdefff80d8799f41b0ffffd87a9fd8799fd8799f41a0ff00ffffff)))))) (\i0 -> \i0 -> i3 (\i0 -> \i0 -> force (i6 i1 (delay i7) (delay ((\i0 -> i5 (i15 (i13 i2)) (\i0 -> i10 (mkPairData (i11 (i14 i3)) i1) i2) (delay i1)) (i2 (i11 i1)))))) i1)) (\i0 -> (\i0 -> i2 (\i0 -> i2 i2 i1)) (\i0 -> i2 (\i0 -> i2 i2 i1)))) (force (force chooseList))) [  ]) (force mkCons)) (force (force fstPair))) (force ifThenElse)) (force tailList)) (force headList)) (\i0 -> i2 (unConstrData i1))) (force (force sndPair))))
+get.sym (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> unBData ((\i0 -> i7 (i10 i1)) ((\i0 -> (\i0 -> i3 (\i0 -> i2 (unMapData i1) (\i0 -> (\i0 -> \i0 -> \i0 -> i2 i3) (mapData i1)) (delay (\i0 -> \i0 -> force i1)))) (\i0 -> (\i0 -> force (i10 (force nullList i1) (delay (\i0 -> \i0 -> force i1)) (delay (\i0 -> \i0 -> i2 i3)))) (i3 (\i0 -> force (i10 (equalsData i1 (iData 0)) (delay (\i0 -> \i0 -> force i1)) (delay (\i0 -> \i0 -> i2 i3)))) i1)) i1) (unMapData ((\i0 -> i10 ((\i0 -> i10 (i10 (i10 i1))) (i11 i1))) ((\i0 -> i10 (i11 i1)) #d8799fd8799f9fd8799fd8799fd8799f41a0ff00ffd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffff80a0a141c0a149736f6d65746f6b656e018080d8799fd8799fd87980d87a80ffd8799fd87b80d87a80ffff9f45ab01fe235c4312301443abcdefff80d8799f41b0ffffd87a9fd8799fd8799f41a0ff00ffffff)))))) (\i0 -> \i0 -> i3 (\i0 -> \i0 -> force (i6 i1 (delay i7) (delay ((\i0 -> i5 (i15 (i13 i2)) (\i0 -> i10 (mkPairData (i11 (i14 i3)) i1) i2) (delay i1)) (i2 (i11 i1)))))) i1)) (\i0 -> (\i0 -> i2 (\i0 -> i2 i2 i1)) (\i0 -> i2 (\i0 -> i2 i2 i1)))) (force (force chooseList))) [  ]) (force mkCons)) (force (force fstPair))) (force ifThenElse)) (force tailList)) (force headList)) (\i0 -> i2 (unConstrData i1))) (force (force sndPair))))
 ScriptPurpose.literal (program 1.0.0 #d8799f58201111111111111111111111111111111111111111111111111111111111111111ff)
 ScriptPurpose.decode (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> force (force ifThenElse (equalsInteger 0 i2) (delay i1) (delay error))) (force (force sndPair) i2)) (force (force fstPair) i1)) (unConstrData #d8799f58201111111111111111111111111111111111111111111111111111111111111111ff)))
