+++ plutarch-test/goldens/api.map.uplc.eval.golden
@@ -1,8 +1,7 @@
-lookup.itself (program 1.0.0 (\i0 -> (\i0 -> \i0 -> \i0 -> (\i0 -> (\i0 -> i2 (\i0 -> i2 i2 i1)) (\i0 -> i2 (\i0 -> i2 i2 i1))) (\i0 -> \i0 -> force (force (force chooseList) i1 (delay (\i0 -> \i0 -> force i1)) (delay (force (force ifThenElse (equalsData (force (force fstPair) (force headList i1)) i4) (delay (i5 (force headList i1))) (delay (i2 (force tailList i1)))))))) i1) (\i0 -> \i0 -> \i0 -> i2 (unIData (force (force sndPair) i3))) (bData i1)))
-lookup.hit (program 1.0.0 (\i0 -> \i0 -> i2 (unIData (force (force sndPair) ( #436b6579
-, #182a )))))
+lookup.itself (program 1.0.0 (\i0 -> (\i0 -> \i0 -> \i0 -> (\i0 -> (\i0 -> i2 (\i0 -> i2 i2 i1)) (\i0 -> i2 (\i0 -> i2 i2 i1))) (\i0 -> \i0 -> force (force (force chooseList) i1 (delay (\i0 -> \i0 -> force i1)) (delay (force (force ifThenElse (equalsData (force (force fstPair) (force headList i1)) i4) (delay (i5 (force headList i1))) (delay (i2 (force tailList i1)))))))) i1) (\i0 -> (\i0 -> \i0 -> \i0 -> i2 i3) (unIData (force (force sndPair) i1))) (bData i1)))
+lookup.hit (program 1.0.0 (\i0 -> \i0 -> i2 42))
 lookup.miss (program 1.0.0 (\i0 -> \i0 -> force i1))
-lookupData.hit (program 1.0.0 (\i0 -> \i0 -> i2 (force (force sndPair) (#436b6579, #182a))))
+lookupData.hit (program 1.0.0 (\i0 -> \i0 -> i2 #182a))
 lookupData.miss (program 1.0.0 (\i0 -> \i0 -> force i1))
 findWithDefault.itself (program 1.0.0 (\i0 -> \i0 -> (\i0 -> \i0 -> \i0 -> \i0 -> (\i0 -> (\i0 -> i2 (\i0 -> i2 i2 i1)) (\i0 -> i2 (\i0 -> i2 i2 i1))) (\i0 -> \i0 -> force (force (force chooseList) i1 (delay i5) (delay (force (force ifThenElse (equalsData (force (force fstPair) (force headList i1)) i6) (delay (i4 (force (force sndPair) (force headList i1)))) (delay (i2 (force tailList i1)))))))) i1) (bData i1) i2 unIData))
 findWithDefault.hit (program 1.0.0 42)
