+++ plutarch-test/goldens/extra.api.dev=true.bench.golden
@@ -1,3 +1,3 @@
-pfindOwnInput {"exBudgetCPU":5917281,"exBudgetMemory":14580,"scriptSizeBytes":319}
-pgetContinuingOutputs {"exBudgetCPU":12048411,"exBudgetMemory":30186,"scriptSizeBytes":444}
-pparseDatum {"exBudgetCPU":4634894,"exBudgetMemory":14706,"scriptSizeBytes":373}
+pfindOwnInput {"exBudgetCPU":6900718,"exBudgetMemory":17404,"scriptSizeBytes":332}
+pgetContinuingOutputs {"exBudgetCPU":13974787,"exBudgetMemory":35970,"scriptSizeBytes":462}
+pparseDatum {"exBudgetCPU":18732961,"exBudgetMemory":59226,"scriptSizeBytes":378}
