+++ plutarch-test/goldens/extra.api.dev=true.uplc.eval.golden
@@ -1,3 +1,3 @@
-pfindOwnInput (program 1.0.0 (\i0 -> \i0 -> i2 (force headList [ #d8799fd8799fd8799f41a0ff00ffd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffff ])))
+pfindOwnInput (program 1.0.0 (\i0 -> \i0 -> i2 #d8799fd8799fd8799f41a0ff00ffd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffff))
 pgetContinuingOutputs (program 1.0.0 [#d8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffff])
-pparseDatum (program 1.0.0 (\i0 -> \i0 -> i2 ((\i0 -> force headList (force tailList ((\i0 -> force (force sndPair) (unConstrData i1)) #d8799f41d09f0102030405060708090affff))) ((\i0 -> (\i0 -> (\i0 -> i2 (\i0 -> i2 i2 i1)) (\i0 -> i2 (\i0 -> i2 i2 i1))) (\i0 -> \i0 -> force (force (force chooseList) i1 (delay [  ]) (delay (force mkCons (i3 (force headList i1)) (i2 (force tailList i1))))))) (\i0 -> (\i0 -> i2) (unIData i1)) (unListData (force headList (force tailList ((\i0 -> force (force sndPair) (unConstrData i1)) #d8799f41d09f0102030405060708090affff))))))))
+pparseDatum (program 1.0.0 (\i0 -> \i0 -> i2 #9f0102030405060708090aff))
