+++ plutarch-test/goldens/extra.api.dev=true.uplc.golden
@@ -1,3 +1,3 @@
-pfindOwnInput (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> force (i5 (equalsInteger 1 i2) (delay ((\i0 -> \i0 -> (\i0 -> (\i0 -> i2 (\i0 -> i2 i2 i1)) (\i0 -> i2 (\i0 -> i2 i2 i1))) (\i0 -> \i0 -> \i0 -> force (force (force chooseList) i1 (delay (\i0 -> \i0 -> force i1)) (delay (force (i10 (i2 (i11 i1)) (delay (\i0 -> \i0 -> i2 (i13 i3))) (delay (i3 i2 (i12 i1)))))))) (\i0 -> equalsData i2 (i9 (i11 i1))) i2) ((\i0 -> unListData (i7 (i9 i1))) (i6 i4)) (i6 i1))) (delay error))) (i8 i2)) (force (force fstPair) i1)) (unConstrData (i3 (i4 i1)))) (i4 #d8799fd8799f9fd8799fd8799fd8799f41a0ff00ffd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffff9fd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffa0a141c0a149736f6d65746f6b656e018080d8799fd8799fd87980d87a80ffd8799fd87b80d87a80ffff9f45ab01fe235c4312301443abcdefff9fd8799f41d09f0102030405060708090affffffd8799f41b0ffffd87a9fd8799fd8799f41a0ff00ffffff)) (force ifThenElse)) (force headList)) (force tailList)) (\i0 -> i2 (unConstrData i1))) (force (force sndPair))))
-pgetContinuingOutputs (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> force (i7 (equalsInteger 1 i2) (delay ((\i0 -> \i0 -> \i0 -> i8 (\i0 -> \i0 -> \i0 -> force (i12 i1 (delay (\i0 -> \i0 -> force i1)) (delay (force (i13 (i2 (i15 i1)) (delay (\i0 -> \i0 -> i2 (i17 i3))) (delay (i3 i2 (i16 i1)))))))) (i11 i1) i3 (\i0 -> (\i0 -> i10 (\i0 -> \i0 -> force (i13 i1 (delay [  ]) (delay ((\i0 -> force (i15 (i4 i1) (delay (force mkCons i1 (i3 (i18 i2)))) (delay (i3 (i18 i2))))) (i16 i1)))))) (i12 ((\i0 -> i14 (i16 i1)) (i13 (i14 (i15 i1))))) i3) (delay (force (force trace "can't get any continuing outputs" (delay error))))) ((\i0 -> unListData (i10 (i12 i1))) (i9 i4)) ((\i0 -> unListData (i10 (i11 (i12 i1)))) (i9 i4)) (i9 i1))) (delay error))) (i11 i2)) (force (force fstPair) i1)) (unConstrData (i6 (i7 i1)))) (i7 #d8799fd8799f9fd8799fd8799fd8799f41a0ff00ffd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffff9fd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffa0a141c0a149736f6d65746f6b656e018080d8799fd8799fd87980d87a80ffd8799fd87b80d87a80ffff9f45ab01fe235c4312301443abcdefff9fd8799f41d09f0102030405060708090affffffd8799f41b0ffffd87a9fd8799fd8799f41a0ff00ffffff)) (\i0 -> (\i0 -> i2 (\i0 -> i2 i2 i1)) (\i0 -> i2 (\i0 -> i2 i2 i1)))) (force (force chooseList))) (force ifThenElse)) (\i0 -> \i0 -> equalsData i2 (i3 (i5 i1)))) (force headList)) (force tailList)) (\i0 -> i2 (unConstrData i1))) (force (force sndPair))))
-pparseDatum (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> \i0 -> i3 (\i0 -> \i0 -> \i0 -> force (i7 i1 (delay (\i0 -> \i0 -> force i1)) (delay (force (force ifThenElse (i2 (i9 i1)) (delay (\i0 -> \i0 -> i2 (i11 i3))) (delay (i3 i2 (i8 i1)))))))) (\i0 -> equalsByteString i3 (unBData (i7 (i8 i1)))) i1 (\i0 -> \i0 -> \i0 -> i2 ((\i0 -> i10 (i9 (i11 i4))) ((\i0 -> i7 (\i0 -> \i0 -> force (i10 i1 (delay [  ]) (delay (force mkCons (i3 (i12 i1)) (i2 (i11 i1))))))) (\i0 -> (\i0 -> i2) (unIData i1)) (unListData (i9 (i8 (i10 i3))))))) (delay (\i0 -> \i0 -> force i1))) #d0 ((\i0 -> unListData (i5 ((\i0 -> i5 (i5 (i5 (i5 (i5 (i5 (i5 (i5 i1)))))))) (i6 i1)))) ((\i0 -> i5 (i6 i1)) #d8799fd8799f9fd8799fd8799fd8799f41a0ff00ffd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffff9fd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffa0a141c0a149736f6d65746f6b656e018080d8799fd8799fd87980d87a80ffd8799fd87b80d87a80ffff9f45ab01fe235c4312301443abcdefff9fd8799f41d09f0102030405060708090affffffd8799f41b0ffffd87a9fd8799fd8799f41a0ff00ffffff))) (\i0 -> (\i0 -> i2 (\i0 -> i2 i2 i1)) (\i0 -> i2 (\i0 -> i2 i2 i1)))) (force (force chooseList))) (force tailList)) (force headList)) (\i0 -> i2 (unConstrData i1))) (force (force sndPair))))
+pfindOwnInput (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> force (i5 (equalsInteger 1 i2) (delay ((\i0 -> \i0 -> (\i0 -> (\i0 -> i2 (\i0 -> i2 i2 i1)) (\i0 -> i2 (\i0 -> i2 i2 i1))) (\i0 -> \i0 -> \i0 -> force (force (force chooseList) i1 (delay (\i0 -> \i0 -> force i1)) (delay (force (i10 (i2 (i11 i1)) (delay ((\i0 -> \i0 -> \i0 -> i2 i3) (i11 i1))) (delay (i3 i2 (i12 i1)))))))) (\i0 -> (\i0 -> equalsData (listData (i12 i3)) (listData (i12 i1))) (i9 (i11 i1))) i2) ((\i0 -> unListData (i7 (i9 i1))) (i6 i4)) (i6 i1))) (delay error))) (i8 i2)) (force (force fstPair) i1)) (unConstrData (i3 (i4 i1)))) (i4 #d8799fd8799f9fd8799fd8799fd8799f41a0ff00ffd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffff9fd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffa0a141c0a149736f6d65746f6b656e018080d8799fd8799fd87980d87a80ffd8799fd87b80d87a80ffff9f45ab01fe235c4312301443abcdefff9fd8799f41d09f0102030405060708090affffffd8799f41b0ffffd87a9fd8799fd8799f41a0ff00ffffff)) (force ifThenElse)) (force headList)) (force tailList)) (\i0 -> i2 (unConstrData i1))) (force (force sndPair))))
+pgetContinuingOutputs (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> force (i7 (equalsInteger 1 i2) (delay ((\i0 -> \i0 -> \i0 -> i8 (\i0 -> \i0 -> \i0 -> force (i12 i1 (delay (\i0 -> \i0 -> force i1)) (delay (force (i13 (i2 (i16 i1)) (delay ((\i0 -> \i0 -> \i0 -> i2 i3) (i16 i1))) (delay (i3 i2 (i17 i1)))))))) (i11 i1) i3 (\i0 -> (\i0 -> i10 (\i0 -> \i0 -> force (i13 i1 (delay [  ]) (delay ((\i0 -> force (i15 (i4 i1) (delay (force mkCons i1 (i3 (i19 i2)))) (delay (i3 (i19 i2))))) (i17 i1)))))) (i12 ((\i0 -> i15 (i17 i1)) (i14 (i15 (i16 i1))))) i3) (delay (force (force trace "can't get any continuing outputs" (delay error))))) ((\i0 -> unListData (i11 (i13 i1))) (i10 i4)) ((\i0 -> unListData (i11 (i12 (i13 i1)))) (i10 i4)) (i10 i1))) (delay error))) (i12 i2)) (force (force fstPair) i1)) (unConstrData (i7 (i8 i1)))) (i8 #d8799fd8799f9fd8799fd8799fd8799f41a0ff00ffd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffff9fd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffa0a141c0a149736f6d65746f6b656e018080d8799fd8799fd87980d87a80ffd8799fd87b80d87a80ffff9f45ab01fe235c4312301443abcdefff9fd8799f41d09f0102030405060708090affffffd8799f41b0ffffd87a9fd8799fd8799f41a0ff00ffffff)) (\i0 -> (\i0 -> i2 (\i0 -> i2 i2 i1)) (\i0 -> i2 (\i0 -> i2 i2 i1)))) (force (force chooseList))) (force ifThenElse)) (\i0 -> \i0 -> i3 i2 (i4 (i6 i1)))) (\i0 -> \i0 -> equalsData (listData (i5 i2)) (listData (i5 i1)))) (force headList)) (force tailList)) (\i0 -> i2 (unConstrData i1))) (force (force sndPair))))
+pparseDatum (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> \i0 -> i3 (\i0 -> \i0 -> \i0 -> force (i7 i1 (delay (\i0 -> \i0 -> force i1)) (delay (force (force ifThenElse (i2 (i9 i1)) (delay ((\i0 -> \i0 -> \i0 -> i2 i3) (i9 i1))) (delay (i3 i2 (i8 i1)))))))) (\i0 -> equalsByteString i3 (unBData (i7 (i8 i1)))) i1 (\i0 -> (\i0 -> \i0 -> \i0 -> i2 i3) ((\i0 -> i8 (i7 (i9 i2))) ((\i0 -> i5 (\i0 -> \i0 -> force (i8 i1 (delay [  ]) (delay (force mkCons (i3 (i10 i1)) (i2 (i9 i1))))))) (\i0 -> (\i0 -> i2) (unIData i1)) (unListData (i7 (i6 (i8 i1))))))) (delay (\i0 -> \i0 -> force i1))) #d0 ((\i0 -> unListData (i5 ((\i0 -> i5 (i5 (i5 (i5 (i5 (i5 (i5 (i5 i1)))))))) (i6 i1)))) ((\i0 -> i5 (i6 i1)) #d8799fd8799f9fd8799fd8799fd8799f41a0ff00ffd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffff9fd8799fd8799fd87a9f41a1ffd87a80ffa0d8799f41d0ffffffa0a141c0a149736f6d65746f6b656e018080d8799fd8799fd87980d87a80ffd8799fd87b80d87a80ffff9f45ab01fe235c4312301443abcdefff9fd8799f41d09f0102030405060708090affffffd8799f41b0ffffd87a9fd8799fd8799f41a0ff00ffffff))) (\i0 -> (\i0 -> i2 (\i0 -> i2 i2 i1)) (\i0 -> i2 (\i0 -> i2 i2 i1)))) (force (force chooseList))) (force tailList)) (force headList)) (\i0 -> i2 (unConstrData i1))) (force (force sndPair))))
