+++ plutarch-test/goldens/maybe.bench.golden
@@ -1,6 +1,6 @@
 eq.true.nothing {"exBudgetCPU":529100,"exBudgetMemory":2400,"scriptSizeBytes":34}
-eq.true.just {"exBudgetCPU":876033,"exBudgetMemory":3001,"scriptSizeBytes":38}
-eq.false.nothing-just {"exBudgetCPU":552100,"exBudgetMemory":2500,"scriptSizeBytes":36}
-eq.false.just-just {"exBudgetCPU":876033,"exBudgetMemory":3001,"scriptSizeBytes":38}
+eq.true.just {"exBudgetCPU":1014033,"exBudgetMemory":3601,"scriptSizeBytes":43}
+eq.false.nothing-just {"exBudgetCPU":621100,"exBudgetMemory":2800,"scriptSizeBytes":38}
+eq.false.just-just {"exBudgetCPU":1014033,"exBudgetMemory":3601,"scriptSizeBytes":43}
 pfromJust.nothing {"exBudgetCPU":100,"exBudgetMemory":100,"scriptSizeBytes":14}
-pfromJust.just {"exBudgetCPU":600033,"exBudgetMemory":1801,"scriptSizeBytes":21}
+pfromJust.just {"exBudgetCPU":669033,"exBudgetMemory":2101,"scriptSizeBytes":23}
