+++ plutarch-test/goldens/pair.bench.golden
@@ -1,4 +1,4 @@
-eq.true {"exBudgetCPU":1728579,"exBudgetMemory":5203,"scriptSizeBytes":59}
-eq.false.fst {"exBudgetCPU":1370589,"exBudgetMemory":4802,"scriptSizeBytes":59}
-eq.false.snd {"exBudgetCPU":1659579,"exBudgetMemory":4903,"scriptSizeBytes":64}
-eq.false.both {"exBudgetCPU":1301589,"exBudgetMemory":4502,"scriptSizeBytes":64}
+eq.true {"exBudgetCPU":1866579,"exBudgetMemory":5803,"scriptSizeBytes":64}
+eq.false.fst {"exBudgetCPU":1508589,"exBudgetMemory":5402,"scriptSizeBytes":64}
+eq.false.snd {"exBudgetCPU":1797579,"exBudgetMemory":5503,"scriptSizeBytes":69}
+eq.false.both {"exBudgetCPU":1439589,"exBudgetMemory":5102,"scriptSizeBytes":69}
