+++ plutarch-test/goldens/pair.uplc.golden
@@ -1,4 +1,4 @@
-eq.true (program 1.0.0 ((\i0 -> (\i0 -> \i0 -> i2 (\i0 -> \i0 -> i3 (\i0 -> \i0 -> force ((\i0 -> \i0 -> force ifThenElse i2 i1 (delay False)) (equalsInteger i4 i2) (delay (equalsString i3 i1)))))) (\i0 -> i1 42 i2) (\i0 -> i1 42 i2)) "Hello"))
-eq.false.fst (program 1.0.0 ((\i0 -> (\i0 -> \i0 -> i2 (\i0 -> \i0 -> i3 (\i0 -> \i0 -> force ((\i0 -> \i0 -> force ifThenElse i2 i1 (delay False)) (equalsInteger i4 i2) (delay (equalsString i3 i1)))))) (\i0 -> i1 42 i2) (\i0 -> i1 24 i2)) "Hello"))
-eq.false.snd (program 1.0.0 ((\i0 -> \i0 -> i2 (\i0 -> \i0 -> i3 (\i0 -> \i0 -> force ((\i0 -> \i0 -> force ifThenElse i2 i1 (delay False)) (equalsInteger i4 i2) (delay (equalsString i3 i1)))))) (\i0 -> i1 42 "Hello") (\i0 -> i1 42 "World")))
-eq.false.both (program 1.0.0 ((\i0 -> \i0 -> i2 (\i0 -> \i0 -> i3 (\i0 -> \i0 -> force ((\i0 -> \i0 -> force ifThenElse i2 i1 (delay False)) (equalsInteger i4 i2) (delay (equalsString i3 i1)))))) (\i0 -> i1 42 "Hello") (\i0 -> i1 24 "World")))
+eq.true (program 1.0.0 ((\i0 -> (\i0 -> \i0 -> i2 (\i0 -> \i0 -> i3 (\i0 -> \i0 -> force ((\i0 -> \i0 -> force ifThenElse i2 i1 (delay False)) (equalsInteger i4 i2) (delay (equalsString i3 i1)))))) ((\i0 -> \i0 -> i1 i2 i3) 42) ((\i0 -> \i0 -> i1 i2 i3) 42)) "Hello"))
+eq.false.fst (program 1.0.0 ((\i0 -> (\i0 -> \i0 -> i2 (\i0 -> \i0 -> i3 (\i0 -> \i0 -> force ((\i0 -> \i0 -> force ifThenElse i2 i1 (delay False)) (equalsInteger i4 i2) (delay (equalsString i3 i1)))))) ((\i0 -> \i0 -> i1 i2 i3) 42) ((\i0 -> \i0 -> i1 i2 i3) 24)) "Hello"))
+eq.false.snd (program 1.0.0 ((\i0 -> \i0 -> i2 (\i0 -> \i0 -> i3 (\i0 -> \i0 -> force ((\i0 -> \i0 -> force ifThenElse i2 i1 (delay False)) (equalsInteger i4 i2) (delay (equalsString i3 i1)))))) ((\i0 -> \i0 -> i1 i2 "Hello") 42) ((\i0 -> \i0 -> i1 i2 "World") 42)))
+eq.false.both (program 1.0.0 ((\i0 -> \i0 -> i2 (\i0 -> \i0 -> i3 (\i0 -> \i0 -> force ((\i0 -> \i0 -> force ifThenElse i2 i1 (delay False)) (equalsInteger i4 i2) (delay (equalsString i3 i1)))))) ((\i0 -> \i0 -> i1 i2 "Hello") 42) ((\i0 -> \i0 -> i1 i2 "World") 24)))
