+++ plutarch-test/goldens/plutustype.deconstr.bench.golden
@@ -33,5 +33,5 @@ matching.raw.sumtype(exhaustive)(ignore-fields).rewarding {"exBudgetCPU":1635210
 matching.raw.sumtype(exhaustive)(ignore-fields).certifying {"exBudgetCPU":2200699,"exBudgetMemory":5470,"scriptSizeBytes":77}
 fields.typed.extract-single {"exBudgetCPU":460976,"exBudgetMemory":1496,"scriptSizeBytes":32}
 fields.raw.extract-single {"exBudgetCPU":460976,"exBudgetMemory":1496,"scriptSizeBytes":32}
-combined.typed.toValidatorHash {"exBudgetCPU":1777528,"exBudgetMemory":5194,"scriptSizeBytes":72}
-combined.raw.toValidatorHash {"exBudgetCPU":1507597,"exBudgetMemory":4362,"scriptSizeBytes":67}
+combined.typed.toValidatorHash {"exBudgetCPU":2012997,"exBudgetMemory":5958,"scriptSizeBytes":75}
+combined.raw.toValidatorHash {"exBudgetCPU":1874997,"exBudgetMemory":5358,"scriptSizeBytes":70}
