+++ plutarch-test/goldens/plutustype.deconstr.uplc.golden
@@ -33,5 +33,5 @@ matching.raw.sumtype(exhaustive)(ignore-fields).rewarding (program 1.0.0 ((\i0 -
 matching.raw.sumtype(exhaustive)(ignore-fields).certifying (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> force (i3 (equalsInteger i1 1) (delay #02) (delay (force (i3 (equalsInteger i1 2) (delay #03) (delay (force (i3 (equalsInteger i1 3) (delay #04) (delay #01))))))))) (force (force fstPair) (unConstrData i1))) #d87c9fd87e80ff) (force ifThenElse)))
 fields.typed.extract-single (program 1.0.0 ((\i0 -> force headList (force (force sndPair) (unConstrData i1))) #d8799fd87a9f41abffd87a80ff))
 fields.raw.extract-single (program 1.0.0 ((\i0 -> force headList (force (force sndPair) (unConstrData i1))) #d8799fd87a9f41abffd87a80ff))
-combined.typed.toValidatorHash (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> force (force ifThenElse (equalsInteger 1 i2) (delay (\i0 -> \i0 -> i2 (unBData (i7 i3)))) (delay (\i0 -> \i0 -> force i1)))) (i5 i2)) (force (force fstPair) i1)) (unConstrData (i2 (i3 (unConstrData i1))))) #d8799fd87a9f41abffd87a80ff) (force headList)) (force (force sndPair))))
-combined.raw.toValidatorHash (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> (\i0 -> force (force ifThenElse (equalsInteger (force (force fstPair) i1) 0) (delay (\i0 -> \i0 -> force i1)) (delay (\i0 -> \i0 -> i2 (unBData (i5 (i6 i3))))))) (unConstrData (i2 (i3 (unConstrData i1))))) #d8799fd87a9f41abffd87a80ff) (force headList)) (force (force sndPair))))
+combined.typed.toValidatorHash (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> (\i0 -> force (force ifThenElse (equalsInteger 1 i2) (delay ((\i0 -> \i0 -> \i0 -> i2 i3) (unBData (i5 i1)))) (delay (\i0 -> \i0 -> force i1)))) (i5 i2)) (force (force fstPair) i1)) (unConstrData (i2 (i3 (unConstrData i1))))) #d8799fd87a9f41abffd87a80ff) (force headList)) (force (force sndPair))))
+combined.raw.toValidatorHash (program 1.0.0 ((\i0 -> (\i0 -> (\i0 -> (\i0 -> force (force ifThenElse (equalsInteger (force (force fstPair) i1) 0) (delay (\i0 -> \i0 -> force i1)) (delay ((\i0 -> \i0 -> \i0 -> i2 i3) (unBData (i3 (i4 i1))))))) (unConstrData (i2 (i3 (unConstrData i1))))) #d8799fd87a9f41abffd87a80ff) (force headList)) (force (force sndPair))))
