+++ plutarch-test/goldens/rational.dev=true.uplc.eval.golden
@@ -1,9 +1,9 @@
 literal (program 1.0.0 (\i0 -> i1 1 2))
-ops.+ (program 1.0.0 (\i0 -> i1 (multiplyInteger 1 (divideInteger 4 4)) (multiplyInteger 1 (divideInteger 4 4))))
-ops.- (program 1.0.0 (\i0 -> i1 (multiplyInteger 1 (divideInteger 1 1)) (multiplyInteger 1 (divideInteger 6 1))))
-ops.* (program 1.0.0 (\i0 -> i1 (multiplyInteger 1 (divideInteger 1 1)) (multiplyInteger 1 (divideInteger 4 1))))
-ops.harmonic-sum (program 1.0.0 (\i0 -> i1 (multiplyInteger 1 (divideInteger 77 1)) (multiplyInteger 1 (divideInteger 60 1))))
-ops.multi-product (program 1.0.0 (\i0 -> i1 (multiplyInteger 1 (divideInteger 1 1)) (multiplyInteger 1 (divideInteger 6 1))))
+ops.+ (program 1.0.0 (\i0 -> i1 1 1))
+ops.- (program 1.0.0 (\i0 -> i1 1 6))
+ops.* (program 1.0.0 (\i0 -> i1 1 4))
+ops.harmonic-sum (program 1.0.0 (\i0 -> i1 77 60))
+ops.multi-product (program 1.0.0 (\i0 -> i1 1 6))
 compare (program 1.0.0 True)
 round.5/3 (program 1.0.0 2)
 round.4/3 (program 1.0.0 1)
@@ -18,7 +18,7 @@ properFraction.-3/2 (program 1.0.0 True)
 properFraction.-4/3 (program 1.0.0 True)
 data.id.0.5 (program 1.0.0 (\i0 -> i1 1 2))
 data.id.2 (program 1.0.0 (\i0 -> i1 2 1))
-data.id.11/3 (program 1.0.0 (\i0 -> i1 (multiplyInteger 1 (divideInteger 11 1)) (multiplyInteger 1 (divideInteger 3 1))))
+data.id.11/3 (program 1.0.0 (\i0 -> i1 11 3))
 div by 0.1/0 (program 1.0.0 error)
 div by 0.recip 0 (program 1.0.0 error)
 div by 0.1/(1-1) (program 1.0.0 error)
