+++ plutarch-test/goldens/show.bench.golden
@@ -1,15 +1,15 @@
-unit {"exBudgetCPU":23100,"exBudgetMemory":200,"scriptSizeBytes":11}
+unit {"exBudgetCPU":92100,"exBudgetMemory":500,"scriptSizeBytes":13}
 bool.true {"exBudgetCPU":379656,"exBudgetMemory":1401,"scriptSizeBytes":33}
 bool.false {"exBudgetCPU":379656,"exBudgetMemory":1401,"scriptSizeBytes":33}
-int.0 {"exBudgetCPU":5982998,"exBudgetMemory":14021,"scriptSizeBytes":317}
-int.5 {"exBudgetCPU":8810443,"exBudgetMemory":20031,"scriptSizeBytes":317}
-int.-5 {"exBudgetCPU":9431574,"exBudgetMemory":20836,"scriptSizeBytes":317}
-int.10 {"exBudgetCPU":11891916,"exBudgetMemory":26051,"scriptSizeBytes":317}
-int.-10 {"exBudgetCPU":12513047,"exBudgetMemory":26856,"scriptSizeBytes":317}
-int.14 {"exBudgetCPU":14153872,"exBudgetMemory":30859,"scriptSizeBytes":317}
-int.-14 {"exBudgetCPU":14775003,"exBudgetMemory":31664,"scriptSizeBytes":317}
-int.102 {"exBudgetCPU":18414677,"exBudgetMemory":39285,"scriptSizeBytes":318}
-int.-102 {"exBudgetCPU":19035808,"exBudgetMemory":40090,"scriptSizeBytes":318}
+int.0 {"exBudgetCPU":6396998,"exBudgetMemory":15821,"scriptSizeBytes":306}
+int.5 {"exBudgetCPU":9224443,"exBudgetMemory":21831,"scriptSizeBytes":306}
+int.-5 {"exBudgetCPU":9983574,"exBudgetMemory":23236,"scriptSizeBytes":306}
+int.10 {"exBudgetCPU":12581916,"exBudgetMemory":29051,"scriptSizeBytes":306}
+int.-10 {"exBudgetCPU":13341047,"exBudgetMemory":30456,"scriptSizeBytes":306}
+int.14 {"exBudgetCPU":14843872,"exBudgetMemory":33859,"scriptSizeBytes":306}
+int.-14 {"exBudgetCPU":15603003,"exBudgetMemory":35264,"scriptSizeBytes":306}
+int.102 {"exBudgetCPU":19380677,"exBudgetMemory":43485,"scriptSizeBytes":307}
+int.-102 {"exBudgetCPU":20139808,"exBudgetMemory":44890,"scriptSizeBytes":307}
 bytestring.empty {"exBudgetCPU":2593943,"exBudgetMemory":9918,"scriptSizeBytes":414}
 bytestring.1 {"exBudgetCPU":10473299,"exBudgetMemory":27270,"scriptSizeBytes":416}
 bytestring.2 {"exBudgetCPU":29710789,"exBudgetMemory":68664,"scriptSizeBytes":417}
@@ -22,14 +22,14 @@ str.quoted {"exBudgetCPU":30518590,"exBudgetMemory":72910,"scriptSizeBytes":138}
 str.slash {"exBudgetCPU":23919293,"exBudgetMemory":57643,"scriptSizeBytes":136}
 str.unicode {"exBudgetCPU":33189676,"exBudgetMemory":79932,"scriptSizeBytes":139}
 str.unicode-quoted {"exBudgetCPU":40139276,"exBudgetMemory":95604,"scriptSizeBytes":141}
-maybe.nothing {"exBudgetCPU":1127100,"exBudgetMemory":5000,"scriptSizeBytes":354}
-maybe.just {"exBudgetCPU":15447308,"exBudgetMemory":34080,"scriptSizeBytes":356}
-either.right {"exBudgetCPU":15540485,"exBudgetMemory":34381,"scriptSizeBytes":370}
-maybe.either {"exBudgetCPU":17345450,"exBudgetMemory":37542,"scriptSizeBytes":431}
-list.nil {"exBudgetCPU":2825985,"exBudgetMemory":11343,"scriptSizeBytes":410}
-list.1 {"exBudgetCPU":9659511,"exBudgetMemory":26464,"scriptSizeBytes":411}
-list.1,2,3 {"exBudgetCPU":27175986,"exBudgetMemory":66756,"scriptSizeBytes":413}
-builtinlist.nil {"exBudgetCPU":2618985,"exBudgetMemory":10443,"scriptSizeBytes":401}
-builtinlist.1,2,3 {"exBudgetCPU":25171058,"exBudgetMemory":56284,"scriptSizeBytes":404}
-pair.int-str {"exBudgetCPU":34106639,"exBudgetMemory":77405,"scriptSizeBytes":464}
-pair.int-list {"exBudgetCPU":40738700,"exBudgetMemory":86414,"scriptSizeBytes":443}
+maybe.nothing {"exBudgetCPU":1265100,"exBudgetMemory":5600,"scriptSizeBytes":342}
+maybe.just {"exBudgetCPU":16206308,"exBudgetMemory":37380,"scriptSizeBytes":347}
+either.right {"exBudgetCPU":16299485,"exBudgetMemory":37681,"scriptSizeBytes":360}
+maybe.either {"exBudgetCPU":18173450,"exBudgetMemory":41142,"scriptSizeBytes":423}
+list.nil {"exBudgetCPU":2963985,"exBudgetMemory":11943,"scriptSizeBytes":397}
+list.1 {"exBudgetCPU":10073511,"exBudgetMemory":28264,"scriptSizeBytes":398}
+list.1,2,3 {"exBudgetCPU":28141986,"exBudgetMemory":70956,"scriptSizeBytes":401}
+builtinlist.nil {"exBudgetCPU":2756985,"exBudgetMemory":11043,"scriptSizeBytes":389}
+builtinlist.1,2,3 {"exBudgetCPU":26137058,"exBudgetMemory":60484,"scriptSizeBytes":392}
+pair.int-str {"exBudgetCPU":34865639,"exBudgetMemory":80705,"scriptSizeBytes":454}
+pair.int-list {"exBudgetCPU":42325700,"exBudgetMemory":93314,"scriptSizeBytes":434}
