+++ plutarch-test/goldens/show.dev=true.bench.golden
@@ -1,1 +1,1 @@
-rational.1/2 {"exBudgetCPU":23554887,"exBudgetMemory":53494,"scriptSizeBytes":595}
+rational.1/2 {"exBudgetCPU":25003887,"exBudgetMemory":59794,"scriptSizeBytes":560}
