+++ plutarch-test/goldens/unit.uplc.golden
@@ -1,5 +1,5 @@
 pcon (program 1.0.0 ())
-pmatch (program 1.0.0 True)
-compare.== (program 1.0.0 True)
-compare.< (program 1.0.0 False)
-compare.<= (program 1.0.0 True)
+pmatch (program 1.0.0 ((\i0 -> True) ()))
+compare.== (program 1.0.0 ((\i0 -> (\i0 -> True) ()) ()))
+compare.< (program 1.0.0 ((\i0 -> (\i0 -> False) ()) ()))
+compare.<= (program 1.0.0 ((\i0 -> (\i0 -> True) ()) ()))
